Prim's Algorithm - Minimum Spanning Tree
This project was for a course in Data Structures and Algorithms. This
program implements Prim's algorithm to find the minimum spanning tree
between cities. The program gives output in two different ways. The
first gives a list with each city, where that city goes, and where the
cities destination also goes. The program also prints out the adjacency
matrix.